Replace function into helper
